MakaPakaApp/MakaPaka/requirements.txt
flask
redis
pyopenssl
flask-jwt-extended 
gunicorn
pyjwt
fpdf